# Draft: Faster devli