@Melirius 2024-11-28
Overflow is possible, need to check.