@mcroomp 2024-11-28
> Overflow is possible, need to check.

((!shifted).wrapping_add(2) + value) as i16

devli_test exhaustively checks all possible combinations (although JPEG shouldn't go beyond 12 bits anyway), but why not?